add keep instrument

+++ poms/portfolios/views.py
@@ -291,7 +291,10 @@ class PortfolioRegisterViewSet(AbstractWithObjectPermissionViewSet):
     def destroy(self, request, *args, **kwargs):
         instance = self.get_object()
 
-        instance.linked_instrument.delete()
+        keep_instrument = request.data.get('keep_instrument')
+
+        if keep_instrument != 'true':
+            instance.linked_instrument.delete()
 
         self.perform_destroy(instance)
         return Response(status=status.HTTP_204_NO_CONTENT)
